Make sure ungit process closes correctly in app.quit
Reason why ungit keeps on running, is because the shell spins up a new process that doe not get killed when the `ungitProcess` gets killed.

However, with this implementation it does not get closed on ctrl-c.
That is because we run it as a detached processes now.
Perhaps better is to use ps-tree. I tried that but could not get it to work.

Current solution: https://azimi.me/2014/12/31/kill-child_process-node-js.html
ps-tree solution: https://www.npmjs.com/package/ps-tree